requirements.txt
streamlit
pandas
numpy
matplotlib
scipy
Pillow
plotly